Bug fix. Remove old transformer implementation.
