--- thumt/models/experimental/__init__.py
@@ -1,2 +1,0 @@
-# coding=utf-8
-# Copyright 2018 The THUMT Authors
